telethon_manager/requirements.txt
telethon
httpx
python-dotenv
python-docx
pdfplumber
